entryCheck

+++ src/main/java/com/truck/service/impl/TransportServiceImpl.java
@@ -281,6 +281,9 @@ public class TransportServiceImpl implements ITransportService {
     public ServerResponse checkEntryByDeclareNum(String declareNum){
         Entry entry = entryMapper.selectByDeclareNum(declareNum);
         if(entry != null){
+            if(Const.EntryStatusEnum.FINISH.getCode() < entry.getStatus()){
+                return ServerResponse.createByErrorMessage("本单已入库，无法替换");
+            }
             entryDetailMapper.deleteByEntryId(entry.getId());
             List<EntryDetail> entryDetailList = entryDetailMapper.selectEntryDetail(entry.getId());
             if(entryDetailList.size() == 0){
@@ -303,10 +306,11 @@ public class TransportServiceImpl implements ITransportService {
             return ServerResponse.createByErrorMessage("系统异常");
         }
         String status = jsonObject.get("status").toString();
+        String msg = jsonObject.get("msg").toString();
         if(status.equals("0")){
             return ServerResponse.createBySuccess();
         }
-        return ServerResponse.createByErrorMessage("清除失败");
+        return ServerResponse.createByErrorMessage(msg);
     }
 
     private long generateEntryNo(){
